+++ packages/web/src/lib/view/index/format/share.ts
