+++ packages/web/src/lib/view/index/random/range/LoadRangeSlider.spec.ts
@@ -8,7 +8,7 @@ import { fireEvent, render } from '@testing-library/svelte'
 import { tick } from 'svelte'
 import { describe, expect, it, vi } from 'vitest'
 
-import { initializeAssembly } from '../../interaction/assembly'
+import { initializeAssembly } from '../../usecase/initialize-assembly'
 
 import LoadRangeSlider from './LoadRangeSlider.test-wrapper.svelte'
