+++ packages/web/src/lib/view/index/usecase/assembly-from-query.spec.ts
@@ -3,7 +3,7 @@ import { latest as regulation } from '$lib/regulation'
 import { assemblyToSearchV2 } from '@ac6_assemble_tool/core/assembly/serialize/as-query-v2'
 import { afterEach, describe, expect, it, vi } from 'vitest'
 
-import { initializeAssembly } from './assembly'
+import { initializeAssembly } from './initialize-assembly'
 import { buildAssemblyFromQuery } from './assembly-from-query'
 
 describe('assembly-from-query', () => {
