+++ packages/web/src/lib/view/index/usecase/assembly-from-query.ts
