+++ packages/web/src/lib/view/index/usecase/derive-parts-pool.ts
